async wrapper with module container
