+++ asyncFunction.js
@@ -1,0 +1,7 @@
+var asyncFunction =  function (func) {
+	return new Promise((resolve) => {
+    			return resolve(func()); 
+  	});
+}
+
+export {asyncFunction}
